The Try class for turning a concrete failure into an abstract failure.
Edward Yang suggested this in the Haskell cafe,
and I agreed it is a good idea.
Specially because c-m-e and (I believe) attempt
already provide something like this to convert
concrete failures to EMT and Attempt respectively.

The implementation does require Functional Deps.
or Type Families. I chose to go with TFs here.

Modified by Michael Snoyman to put new additions in the export list.